requirements.txt
black==23.12.0
evaluate==0.4.1
fastapi==0.109.2
flake8==7.0.0
gradio==4.19.1
gunicorn==22.0.0
huggingface-hub==0.26.3
ipykernel==6.29.0
ipython==8.21.0
ipywidgets==8.1.2
langchain==0.2.0
langchain-core==0.2.0
langchain-community==0.2.1
langchain-huggingface==0.0.1
langchain-openai==0.1.7
loguru==0.7.2
numpy==1.26.3
pandas==2.2.0
pillow==10.2.0
pydantic-settings==2.2.1
pytest==8.2.2
python-dotenv==1.0.1
requests==2.28.1
scikit-learn==1.4.0
scipy==1.12.0
tokenizers==0.19.1
tomli==2.0.1
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.46.1